% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_character.R
\name{first_character}
\alias{first_character}
\title{Return the first character of a string.}
\usage{
first_character(s)
}
\arguments{
\item{s:}{A string}
}
\value{
First character of s
}
\description{
Return the first character of a string.
}
\examples{
first_character("Some String") # returns 'S'
}
\author{
Davis Marc Vigneault
}

