% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.R
\name{format.p.value}
\alias{format.p.value}
\title{Format a `p` value as expected for a table or figure.}
\usage{
\method{format}{p.value}(p, alpha = 0.05)
}
\arguments{
\item{p:}{A `p` value}

\item{alpha:}{Threshold for significance}
}
\value{
A formatted `p` value
}
\description{
Format a `p` value as expected for a table or figure.
}
\examples{
format.p.value(0.5) # returns NS
format.p.value(0.05) # returns 0.5
format.p.value(0.005) # returns <0.01
format.p.value(0.0005) # returns <0.001
format.p.value(0.00005) # returns <0.0001
}
\author{
Davis Marc Vigneault
}

