% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.R
\name{format_p_value}
\alias{format_p_value}
\title{Format a `p` value as expected for a table or figure.}
\usage{
format_p_value(p, alpha = 0.05)
}
\arguments{
\item{p}{A `p` value}

\item{alpha}{Threshold for significance}
}
\value{
A formatted `p` value
}
\description{
Format a `p` value as expected for a table or figure.
}
\examples{
format_p_value(0.5) # returns NS
format_p_value(0.05) # returns 0.5
format_p_value(0.005) # returns <0.01
format_p_value(0.0005) # returns <0.001
format_p_value(0.00005) # returns <0.0001
}
\author{
Davis Marc Vigneault
}

