% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{\%\%}
\alias{\%\%}
\alias{\% \%}
\title{%%}
\usage{
... \%\% NA

x \% \% y
}
\arguments{
\item{...}{}

\item{...}{}
}
\value{
paste together with no space

paste together with space
}
\description{
%%

% %
}
