% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{keep_these}
\alias{keep_these}
\title{keep_these}
\usage{
keep_these(x, keep)
}
\arguments{
\item{x}{a vector}

\item{keep}{a vector of values to keep, exclusively, from \code{x}}
}
\value{
\code{x} with only those elements matching any in \code{remove}
}
\description{
keep_these
}
