% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_character.R
\name{last.character}
\alias{last.character}
\title{Return the last character of a string.}
\usage{
last.character(s)
}
\arguments{
\item{s:}{A string}
}
\value{
Last character of s
}
\description{
Return the last character of a string.
}
\examples{
last.character("Some String") # returns 'g'
}
\author{
Davis Marc Vigneault
}

