% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{library_many}
\alias{library_many}
\title{library_many}
\usage{
library_many(...)
}
\arguments{
\item{x}{character vector of packages to load}
}
\value{
load packages in \code{x} into session/namespace
}
\description{
library_many
}
