% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{most_common}
\alias{most_common}
\title{most_common}
\usage{
most_common(x)
}
\arguments{
\item{x}{vector}
}
\value{
return which value in \code{x} occurs most often. Only 1 returned in case of tie
}
\description{
most_common
}
