% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{most_commons}
\alias{most_commons}
\title{most_commons}
\usage{
most_commons(x)
}
\arguments{
\item{x}{vector}
}
\value{
return which value in \code{x} occurs most often. All returned in case of tie
}
\description{
most_commons
}
