% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{nlapply}
\alias{nlapply}
\title{nlapply}
\usage{
nlapply(n, ...)
}
\arguments{
\item{n}{first argument to lapply function, with names}

\item{...}{}
}
\value{
\code{lapply} call with named list returned
}
\description{
nlapply
}
