% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{remove_intercept}
\alias{remove_intercept}
\title{remove_intercept}
\usage{
remove_intercept(x)
}
\arguments{
\item{x}{matrix or data.frame}
}
\value{
If there is an intercept column (all 1s), it is removed
}
\description{
remove_intercept
}
