% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{remove_last}
\alias{remove_last}
\title{remove_last}
\usage{
remove_last(x, n = 1)
}
\arguments{
\item{x}{vector}

\item{n}{number of tail to remove}
}
\value{
\code{x} with the last \code{n} values removed
}
\description{
remove_last
}
