% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{remove_these}
\alias{remove_these}
\title{remove_these}
\usage{
remove_these(x, remove)
}
\arguments{
\item{x}{a vector}

\item{remove}{a vector of values to remove from \code{x}}
}
\value{
\code{x} with all those elements matching any in \code{remove} removed
}
\description{
remove_these
}
