% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Rotate a vector `v` by `n`.}
\usage{
roll(v, n)
}
\arguments{
\item{v:}{A vector}

\item{n:}{An offset}
}
\value{
A rotated vector.
}
\description{
Rotate a vector `v` by `n`.
}
\examples{
roll(1:5, 2) # returns c(4, 5, 1, 2, 3)
roll(1:5, 7) # returns c(4, 5, 1, 2, 3)
roll(1:5, -2) # returns c(3, 4, 5, 1, 2)
roll(1:5, -7) # returns c(3, 4, 5, 1, 2)
}
\author{
Davis Marc Vigneault
}

